Move serializers in the package hierarchy
